Reinstate original Udmurt mobile keyboard, remove SME desktop layouts
The original keyboard was lost in the kbdgen 1-to-2 conversion, and no-one noticed.

Also, all desktop keyboard data is purely SME, and of no value. Thus removed.

What is left is a plain Android keyboard, that can be used as a basis for iOS keyboards.

Desktop keyboards are left for the future.